"update feat delete token"

+++ internal/service/token.go
@@ -30,7 +30,7 @@ func (s *TokenService) CreateToken(symbol string, price float64) error {
 		return fmt.Errorf("This token was duplicated. ")
 	}
 	if err := s.userRepo.CreateToken(newToken); err != nil {
-		logrus.Errorf("Failed to create new user: %s", err.Error())
+		logrus.Errorf("Failed to create new token: %s", err.Error())
 		return fmt.Errorf("Internal server error. ")
 	}
 	return nil
@@ -48,8 +48,14 @@ func (s *TokenService) DeleteToken(address uuid.UUID) error {
 	newToken := &model.Token{
 		Address: address,
 	}
+
+	if s.ValidateTokenInUse(address) {
+		logrus.Errorf("Failed to delete token. Token InUse. ")
+		return fmt.Errorf("Internal server error. ")
+	}
+
 	if err := s.userRepo.DeleteToken(newToken); err != nil {
-		logrus.Errorf("Failed to create new user: %s", err.Error())
+		logrus.Errorf("Failed to delete token. : %s", err.Error())
 		return fmt.Errorf("Internal server error. ")
 	}
 	return nil
@@ -115,3 +121,44 @@ func (s *TokenService) ValidateToken(address uuid.UUID) bool {
 	}
 	return true
 }
+
+func (s *TokenService) ValidateTokenInUse(address uuid.UUID) bool {
+	if s.userRepo.ValidateTokenInUse(address) {
+	}
+	return true
+}
+
+//func (s *TokenService) AirDropNewUserToken(senderWalletAddress uuid.UUID, receiverWalletAddress uuid.UUID, tokenAddress uuid.UUID, amount float64) error {
+//
+//	newtransaction := &model.Transaction{
+//		FromAddress:  senderWalletAddress,
+//		ToAddress:    receiverWalletAddress,
+//		Amount:       amount,
+//		TokenAddress: tokenAddress,
+//	}
+//
+//	if !s.ValidateWallet(senderWalletAddress) {
+//		logrus.Errorf("Sender wallet not found. ")
+//		return fmt.Errorf("Sender wallet not found. ")
+//	}
+//
+//	if !s.ValidateWallet(receiverWalletAddress) {
+//		logrus.Errorf("Receiver wallet not found. ")
+//		return fmt.Errorf("Receiver wallet not found. ")
+//	}
+//
+//	if !s.ValidateToken(tokenAddress) {
+//		logrus.Errorf("Token not found. ")
+//		return fmt.Errorf("Token not found. ")
+//	}
+//	if newtransaction.Amount < 0 {
+//		fmt.Print("Amount must be larger than 0. ")
+//	}
+//
+//	if err := s.userRepo.SendUserToken(newtransaction); err != nil {
+//		logrus.Errorf("Failed to create new user: %s", err.Error())
+//		return fmt.Errorf("Internal server error. ")
+//	}
+//	return nil
+//
+//}
